Fixed bug in run struct, added logging of run params

+++ include/runs.hpp
@@ -24,7 +24,7 @@ public:
     float decay_rate;
     int initial_proteins;
     float max_mut_float;
-    float max_mut_bits;
+    int max_mut_bits;
     int fitness_log_interval;
 
     int run_index; //index of run in file (this is filled in by the Runs() constructor and is not part of toml file)
